09-09 CSS:media설정 + Student 최종성적 출력 + 좌측 메뉴 Class 이벤트 판정 개선 + 클래스 선택화면 추가

+++ src/main/java/com/example/app/model/dto/response/repository/TotalGradeMapping.java
@@ -8,5 +8,5 @@ public interface TotalGradeMapping {
      Long getTask();
      Long getStudent();
      Long getCount();
-     Long getSum();
+     Double getSum();
 }
